Tests for fetchers and formatters
